Some changes to make the interface a little more consistent.
The --N-RHS options shouldn't be followed by "="; no other option uses
that convention.  Having the symbols be the next argument is how all the
other "--" options work, admittedly not quite the same as -N, etc.

Using -X was probably not really sensible anyway; at least add
a synonym of --constant.